:technologist: Improve crash handling - now works correctly

+++ tools/RegiDevelServer.mjs
@@ -3,6 +3,7 @@ import readline from 'readline'
 import chalk from 'chalk'
 import logger from 'electron-log'
 import { stdin, stdout } from 'node:process'
+
 const log = logger.scope('Regi-DevServer')
 import esbuild from 'esbuild'
 
@@ -24,6 +25,7 @@ async function startPreview(cmd) {
 }
 
 async function main() {
+  let isCrashed = false
   const buildParams = {
     entryPoints: ['./src/Server/regi.js'],
     format: 'cjs',
@@ -35,12 +37,18 @@ async function main() {
     target: 'node16',
     watch: {
       async onRebuild(error, result) {
+        isCrashed = false
         log.info(`${chalk.bgYellowBright.bold('Reloading')} changes detected`)
+
         if (error) {
+          isCrashed = true
           log.error(`${chalk.bgRedBright.bold('Error')} Build failed — ${error}`)
-          await startPreview(previewCmd).then(ps => {
-            previewProcess.ps = ps
-          })
+          await previewProcess.ps.kill('SIGTERM')
+          if (!isCrashed) {
+            await startPreview(previewCmd).then(ps => {
+              previewProcess.ps = ps
+            })
+          }
         } else {
           await previewProcess.ps.kill('SIGTERM')
           await startPreview(previewCmd).then(ps => {
@@ -62,11 +70,15 @@ async function main() {
 
     rl.on('line', async (input) => {
       if (input.match('^rs$')) {
-        await previewProcess.ps.kill('SIGTERM')
-        await startPreview(previewCmd).then(ps => {
-          previewProcess.ps = ps
-          log.info(`${chalk.bgGreenBright.bold('Restarted')} !`)
-        })
+        if (isCrashed) {
+          log.error(`${chalk.bgRedBright.bold('Error')} Please fix errors before reloading`)
+        } else {
+          await previewProcess.ps.kill('SIGTERM')
+          await startPreview(previewCmd).then(ps => {
+            previewProcess.ps = ps
+            log.info(`${chalk.bgGreenBright.bold('Restarted')} !`)
+          })
+        }
       }
     })
